backend/requirements.txt
django
django-extensions
pillow
djangorestframework
djangorestframework-simplejwt